Arreglado bug al terminar proceso no iniciado

+++ edis_c/interfaz/contenedor_secundario/salida_widget.py
@@ -198,7 +198,8 @@ class EjecutarWidget(QWidget):
 
     def terminar_proceso(self):
         """ Termina el proceso """
-
+        if not self.pro:
+            return
         self.pro.terminate()
